requirements.txt
streamlit
pandas
matplotlib
wordcloud
konlpy
transformers
torch
openai
python-dotenv
requests